feat/created services contact

+++ src/modules/information/information.service.ts
@@ -1,18 +1,24 @@
 import {
   ConflictException,
+  ForbiddenException,
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
 import { CreateInformationDto } from './dto/create-information.dto';
 import { UpdateInformationDto } from './dto/update-information.dto';
 import { InformationRepository } from './repositories/information.repository';
-import { ClientRepository } from '../clients/repositories/clients.repository';
+
 
 @Injectable()
 export class InformationService {
   constructor(private informationRepository: InformationRepository) {}
 
-  async create(id: string, createInformationDto: CreateInformationDto) {
+  async create(
+    id: string,
+    createInformationDto: CreateInformationDto,
+    userId: string,
+    typeUser: string
+  ) {
     if (createInformationDto.email) {
       const findEmail = await this.informationRepository.findInforByEmail(
         createInformationDto.email,
@@ -31,9 +37,14 @@ export class InformationService {
       }
     }
 
+    if (typeUser == "client" && id !== userId) {
+      throw new ForbiddenException('Insufficient permission');
+    }
+
     const newInformation = await this.informationRepository.create(
       id,
       createInformationDto,
+      typeUser
     );
     return newInformation;
   }
@@ -46,7 +57,12 @@ export class InformationService {
     return information;
   }
 
-  async update(inforId: string, updateInformationDto: UpdateInformationDto) {
+  async update(
+    inforId: string,
+    updateInformationDto: UpdateInformationDto,
+    userId: string,
+    typeUser: string
+  ) {
     const information = await this.informationRepository.findOne(inforId);
     if (!information) {
       throw new NotFoundException('Informations not found');
@@ -70,6 +86,18 @@ export class InformationService {
       }
     }
 
+    if (typeUser === "client" && information.clientId !== userId) {
+      throw new ForbiddenException('Insufficient permission');
+    }
+
+    if(typeUser === "contact") {
+      const clientIdByContact = await this.informationRepository.findContact(information.contactId)
+
+      if(clientIdByContact !== userId) {
+        throw new ForbiddenException('Insufficient permission');
+      }
+    }
+
     const informationUpdate = this.informationRepository.update(
       inforId,
       updateInformationDto,
@@ -78,10 +106,22 @@ export class InformationService {
     return informationUpdate;
   }
 
-  async remove(inforId: string) {
+  async remove(inforId: string, userId: string, typeUser: string) {
     const information = await this.informationRepository.findOne(inforId);
     if (!information) {
       throw new NotFoundException('Informations not found');
+    }
+
+    if (typeUser === "client" && information.clientId !== userId) {
+      throw new ForbiddenException('Insufficient permission');
+    }
+
+    if(typeUser === "contact") {
+      const clientIdByContact = await this.informationRepository.findContact(information.contactId)
+
+      if(clientIdByContact !== userId) {
+        throw new ForbiddenException('Insufficient permission');
+      }
     }
 
     await this.informationRepository.delete(inforId);
